refactor(agent): co-locate ingress interfaces with their consumers
